Agrego sanitizaciones para disciplinas

+++ admin/src/web/controllers/Disciplina.py
@@ -1,9 +1,10 @@
 from flask import Blueprint, render_template, request,jsonify, redirect, flash
 from src.core.db import db_session
 from src.core.models.Disciplina import Disciplina
+from src.core.models.Categoria import Categoria
 from src.core.models.Socio import Socio
 from src.core.models.relations.SocioSuscriptoDisciplina import SocioSuscriptoDisciplina
-
+from src.web.validators.validatorsDisciplinas import validate_data
 from src.web.controllers.Auth import allowed_request
 from src.web.controllers.FactoryCrud import get_all_docs_json, get_doc_json, create_doc_json, delete_doc_json, exists_entity,get_all_docs_paginated_json
 
@@ -29,7 +30,10 @@ def get_user(id):
 @disciplines_blueprint.route("/create", methods=["POST"])
 def create_discipline():
     disc = request.form.to_dict()
-    print(disc)
+    error = validate_data(disc)
+    if (error):
+        flash(error)
+        return render_template('admin_disciplinas_new.html', categories=get_all_docs_json(Categoria))
     if 'habilitada' in disc.keys():
         disc['habilitada'] = True
     else:
@@ -40,8 +44,16 @@ def create_discipline():
 @disciplines_blueprint.route("/update/<int:id>", methods=["POST"])
 def update_discipline(id):
     disc = request.form.to_dict()
-    print(disc)
     result = db_session.query(Disciplina).filter_by(id = id).all()
+    #Chequeo que no hayan cambiado el action del form
+    if(result == []):
+        errorMsg = "La disciplina que queres editar no existe"
+        flash(errorMsg)
+        return redirect("/admin/disciplines/0")
+    error = validate_data(disc,"UPDATE")
+    if (error):
+        flash(error)
+        return render_template('admin_disciplinas_edit.html', discipline=get_doc_json(Disciplina, id), categories=get_all_docs_json(Categoria))
     if 'habilitada' in disc.keys():
         disc['habilitada'] = True
     else:
@@ -54,13 +66,23 @@ def update_discipline(id):
 
 @disciplines_blueprint.route("/delete/<id>", methods=["DELETE","GET"])
 def delete_discipline(id):
-    delete_doc_json(Disciplina, id)
+    
+    #Chequeo que no hayan cambiado el id a borrar por uno invalido
+    if (not exists_entity(Disciplina,id)):
+        errorMsg = "Error: La disciplina a borrar no existe"
+        flash(errorMsg)
+    else:
+        delete_doc_json(Disciplina, id)
     return redirect("/admin/disciplines/0")
 
 
 @disciplines_blueprint.route("/switch/<int:id>/<string:state>")
 def switch(id,state):
     result = db_session.query(Disciplina).filter_by(id = id).all()
+    if (result == []):
+        errorMsg = "Error: La disciplina no existe"
+        flash(errorMsg)
+        return redirect("/admin/disciplines/0")
     updated_disc = result[0]
     if(state == "true"):
         updated_disc.habilitada = True
